requirements.txt
# 核心依赖
streamlit>=1.28.0
fastapi>=0.104.0
uvicorn>=0.24.0

# 文档处理
pdfplumber>=0.7.4
python-docx>=0.8.11

# 中文处理
jieba>=0.42.1

# 数据库
sqlalchemy>=2.0.0
# sqlite3  # Python内置，无需安装

# LLM集成
openai>=1.3.0
anthropic>=0.18.0

# 工具库
python-multipart>=0.0.6
aiofiles>=23.2.0

# 开发依赖
pytest>=7.4.0
black>=23.0.0

# ==================== 可选依赖 ====================
# 以下依赖是可选的，按需安装

# 高质量文档解析 - MinerU
# 安装方法：pip install mineru
# 或从源码：pip install git+https://github.com/opendatalab/MinerU.git
# mineru>=0.1.0

# 向量数据库 - 用于语义检索
# qdrant-client>=1.6.0
# chromadb>=0.4.0
# faiss-cpu>=1.7.0

# 进阶中文NLP
# pkuseg>=0.0.25
# thulac>=0.2.0

# PDF高级处理（替代pdfplumber）
# pymupdf>=1.23.0

# OCR支持（用于图片/扫描PDF）
# pytesseract>=0.3.10
# Pillow>=10.0.0
